requirements.txt
fastapi
openai
pandas
scikit-learn
python-dotenv
uvicorn
